Remove some unused Result<> return types
Now that the previous commits moved a lot of the actual "read from the
stream"-stuff out of this code, there is actually no way for an error to
occur in lot of functions. Thus, this commit can remove all the
Result<>s from the reading code.

One small complication is poll_check_for_reply_or_error(). This function
sends a sync to the server in some cases, to ensure that a
larger-than-requested sequence number appears. Since this sending can
still fail, it needs a Result<> return. To handle this, the function is
split up into prepare_check_for_reply_or_error() (which sends a sync, if
necessary) and poll_check_for_reply_or_error() (which only polls for
replies).

Signed-off-by: Uli Schlachter <psychon@znc.in>