Add `serialize_into` to the `Serialize` trait.
This functions, instead of returning a vector (or a fixed length
array, appends the serialized structure into an existing vector).

The purpose of this function is to reduce the number of reallocations
when serializing structures with variable length fields.

For example, in `xinput::DeviceClass::serialize` we had:

 ```
 fn serialize(&self) -> Self::Bytes {
     let mut result = Vec::new();
     ...
     result.extend(self.data.serialize());
     ...
 }
 ```

`self.data.serialize()` created a vector, and `extend` appended
that vector to `result` and discarded it.

Now we have:

 ```
 fn serialize_into(&self, bytes: &mut Vec<u8>) {
     ...
     self.data.serialize_into(bytes);
     ...
 }
 ```
`self.data` is serialized directly into the final vector (`bytes`)
without creating an intermediary vector.