Fix "dead-lock" in RustConnection with multiple threads
A single thread that called wait_for_event() would lock the mutex and
then block while waiting for something to be read. This meant that other
threads could not write requests any more, because they would have to
wait for the mutex.

Fix this with a condition variable: A boolean tracks whether some thread
is currently trying to read a packet from the wire. If there is no
reader yet, a thread can become the reader and read a packet. While
reading, it drops the mutex. Another thread that also waits for some
packet would then see that there is already a reader and then wait on
the condition variable that signals that something was read.

This commit only handles the "waiting for some packet and still allow
other threads to sent requests" case. If a thread tries to write to the
socket, it still blocks out other threads, even if they could make
progress (e.g. generate_id()). I believe that writing should "almost
never" block. Let's wait for complaints.

Fixes: https://github.com/psychon/x11rb/issues/203
Signed-off-by: Uli Schlachter <psychon@znc.in>