Correctly serialize lists with expressions
xproto::SetupAuthenticate has a list with a length of length_field*4.
res::ClientIdValue has a list with a length of length_field/4. In both
cases, the code generator just ignored the expression so far. This
commit fixes that and fixes the serialisation of these structs.

This commit also adds a regression test for this bug.

Signed-off-by: Uli Schlachter <psychon@znc.in>