Add RustConnection::for_connected_stream()
This commit adds RustConnection::for_connected_stream(). This function
gets a Read and Write instance and a Setup. The idea is that the streams
represent a raw connection to an X11 server that was just connected and
the given Setup was received.

The intended use of this API is to simplify some tests.

Signed-off-by: Uli Schlachter <psychon@znc.in>