Allow getting full sequence number of events
This commit adds two new methods to the Connection trait that allow to
get an event together with its sequence number.

The relevant code in RustConnection already has the sequence number and
so just has to return it.

The code in XCBConnection can get a 32 bit sequence number, because
libxcb inserts that into events. Getting the high bits of the 64 bit
sequence number is left as an exercise for the future (marked with a new
FIXME).

Fixes: https://github.com/psychon/x11rb/issues/202
Signed-off-by: Uli Schlachter <psychon@znc.in>