Move multithread_test from examples to tests
It is really a test and not an example. However, it needs some X11
server to do its test. Thus, this commit adds some implementation that
parsed the requests that are received and replies to them as needed. I
verified that this is still enough to reproduce the original bug.

Supersedes: https://github.com/psychon/x11rb/pull/233
Signed-off-by: Uli Schlachter <psychon@znc.in>