Fix xkb::SelectEvents
In all of xcb-proto, a <switch> switches on a field directly. All of
xcb-proto? Well, not entirely! One small issue with xkb::SelectEvents
exists. This <switch> uses a combination of three parameters for
switching. Because of how xcb-proto works, the code accidentially
assumed that only the first of these three values is "important" and
switched on that.

This commit fixes the code generator. Instead of assuming that we have a
"trivial" expression, this is now explicitly checked. Since there is
only one non-trivial case to handle, we assert that we are seeing this
case and then do some ugly magic to have the rest of the code generator
continue. This is the change in handle_switches().

Since we can no longer infer the value of the <switch> field, the values
that are present in the Aux structure must now match what is described
by the other fields. For this reason, an assert_eq!() is added to the
generated code of the request function. This is the change in
request_implementation().

Fixes: https://github.com/psychon/x11rb/issues/281
Signed-off-by: Uli Schlachter <psychon@znc.in>