Code generator: Skip write if content is correct
This commit makes the code generator not overwrite its output files with
content that they already have.

The effect of this is that the time of last modification of the file
does not change, thus cargo does not rebuild the code. This has a lot
impact on build times. The idea is to make working on the code generator
less time consuming.

Signed-off-by: Uli Schlachter <psychon@znc.in>