Use enums where it makes sense
As an example, I will use VISUALTYPE's class member. It is defined as:

    <field type="CARD8" name="class" enum="VisualClass" />

Up to now, this resulted in a field 'pub class: u8' in struct
Visualtype. The enum was completely ignored.

This commit changes the code generator to take into account this enum
information. Thus, the field is now 'pub class: VisualClass'.

Doing so mainly requires to fix all the serialisation and parsing code
to transform things as necessary.

This might be the "most breaking" change that was done on x11rb so
far...

Fixes: https://github.com/psychon/x11rb/issues/263
Signed-off-by: Uli Schlachter <psychon@znc.in>