Add a test for a bug in ExtensionInformation
When getting the reply to a QueryExtension request fails
(Cookie::reply() returns an Err), ExtensionInformation does not update
its internal state. Thus, if later the same extension is queried again,
the internal state would still say that a QueryExtension request was
sent, but its reply was not yet collected. Thus, this would ask for the
reply again. In the current implementation, such a "re-query" would just
hang.

This commit adds a test case for this behaviour.

Signed-off-by: Uli Schlachter <psychon@znc.in>