Simplify parsing code
This commit continues an idea originally attempted by @dalcde. Instead
of generating the following code:

    let (response_type, new_remaining) = u8::try_parse(remaining)?;
    remaining = new_remaining;

...this code is generated:

    let (response_type, remaining) = u8::try_parse(remaining)?;

This small looking change results in lots of saved lines:

 32 files changed, 7061 insertions(+), 12964 deletions(-)

A downside of this is that loops become a bit more complicated (because
they have to keep the old form and thus need an extra 'let mut remaining
= remaining;') and parsing switches also becomes a bit more complicated
(because there the 'remaining' variable also needs to be passed out of
its enclosing scope to an outer one). However, all in all, this is still
a net benefit and makes a lot of the generated code a lot easier to
read.

Thanks to @dalcde for this great idea.

Fixes: https://github.com/psychon/x11rb/issues/289
Signed-off-by: Uli Schlachter <psychon@znc.in>