Make property and type arguments of ChangeProperty generic
Before this, the arguments 'property' and 'type' of ChangeProperty had
type u32. This commit changes that into (basically) impl Into<ATOM>.
That way, for example xproto::ATOM::WM_NAME can be passed in directly.

This is a breaking change, because code that previously passed e.g.
Atom::WM_NAME.into() to change_property no longer compiles (because the
compiler cannot infer which type the argument should be converted to).

Fixes: https://github.com/psychon/x11rb/issues/21
Signed-off-by: Uli Schlachter <psychon@znc.in>