Use proper type names
Up to now, xproto.rs contained this:

  pub type WINDOW = u32;

This type name does not follow the usual Rust naming convention. The
reason for this is that xproto also has an enum with name Window and the
two names collided (=were identical).

This commit changes the type alias to be called Window. The name
collision is handled by appending 'Enum' to the name of enums that
collide with type aliases.

There is the following uglyness in this commit:

  if field_type.is_simple and field_type.__class__.__name__ != "Enum":

Suggestions on how to do this differently / less ugly are welcome. It
seems like xcb-proto does not want one to easily tell apart enums and
"simple" types...

Fixes: https://github.com/psychon/x11rb/issues/13
Signed-off-by: Uli Schlachter <psychon@znc.in>