Change the Connection trait to provide parsed events
This changes the event functions in the Connection trait to return Event
instead of GenericEvent. The old API is retained with names containing
"raw_event".

Implementations only have to implement the "raw_event" API and the other
functions are provided. They use the parse_event() API to do the actual
parsing.

Signed-off-by: Uli Schlachter <psychon@znc.in>