XCBConnection: Properly reconstruct sequence numbers
Events only come with a 32 bit sequence number from libxcb. We want to
use 64 bit sequence numbers everywhere. So far, this was handled by a
simply conversion, which means that the high bits were always zero.

This commit uses a recent sequence number to reconstruct the high bits
instead.

The old code worked fine the first 4-billion-or-so times. This new code
works fine as long as we see a new sequence number every 2-billion-or-so
times. That should be pretty much always.

Signed-off-by: Uli Schlachter <psychon@znc.in>