Generated code: Simplify request length calculation
Instead of trying to compute the request length up front, the request
length is now first set to zero. The correct value is then only written
later. This simplifies things, since we no longer have to know the
request length at the beginning of the function.

This idea is originally from @eduardosm, thanks!

Signed-off-by: Uli Schlachter <psychon@znc.in>