Add a mergify configuration
This configuration makes mergify merge PRs that passed CI and that have
one approved review. PRs that are marked with the no-mergify label are
ignored.

When the conditions are fulfilled, mergify merges the PR and deletes the
branch that was merged (if it is in the same repo).

Oh and old reviews are automatically dismissed when new commits are
pushed.

Signed-off-by: Uli Schlachter <psychon@znc.in>