Panic on some length overflows
If a list with an u8 length field should have 300 elements, we have a
problem. Before this commit, the code generator just truncated the
length to an u8 for the length field, but then also serialised the full
list. This means it generated a broken serialisation.

This commit causes a panic instead that says which field has too many
elements.

Related-to: https://github.com/psychon/x11rb/issues/352
Signed-off-by: Uli Schlachter <psychon@znc.in>