Use enums for aux structs where possible
The 'backing_store' field of CreateWindowAux describes, well, the
backing store of the window that should be created. The possible values
are described by the BackingStore enum. However, up to now,
CreateWindowAux still required an u32 argument for this.

This commit changes CreateWindowAux to instead use the BackingStore
enum.

This change was originally thought of by @eduardosm, thanks!

Signed-off-by: Uli Schlachter <psychon@znc.in>