Do not cast bools to u8 before serialising
We implement the Serialize trait for bool for quite a while now, so we
can just use that.

The only complication is xinput's GrabOwner enum. This enum has just two
values and appears in a <field> with type="BOOL". Since the enum is not
convertible to bool, we have to convert it to u8 in this case.

Signed-off-by: Uli Schlachter <psychon@znc.in>