Remove the mock connection in IDAllocator test
Instead of creating a mock of a RequestConnection to test this ID
allocator, this commit changes the test to instead use a callback
function. This allows to get rid of about 100 lines of code. Plus, the
test becomes more precise: It now specifies exactly the place where the
code should send a GetXIDRange request.

Signed-off-by: Uli Schlachter <psychon@znc.in>