Check arithmetic and conversion in generated expressions
Conversions now use `try_from` and arithmetic operations now use `checked_*` functions.

In parse code, failures are propagated to the caller with a `ParseError`. In serialize code, failures produce a panic.

Additionally, expressions now always use `u32` for intermediate values. All fields involved in expressions are `CARD32` or smaller, so a lot of `T::try_from` are now `u32::from`. For list lengths, a `usize::try_from` is used to convert the final value of the expression.