Remove the OsString medling find_cursor
Reading a bunch of bytes from a file and interpreting them as an OsStr
is easy on Unix. On Windows... not so much. This commits cleans things
up by just requiring valid utf8 in the file instead. That way, Windows
is no longer a special case and everything becomes easier.

Signed-off-by: Uli Schlachter <psychon@znc.in>