ScreenSaver: Use Gravity enum instead of u32
Using the Gravity enum instead of u32 is possible in the ScreenSaver
extension because this enum is only sent to the server, but never
received, so it only needs to be serialised. For this, it is not a
problem that the enum has two variants that are represented by the same
value.

Doing this actually simplifies the code in the generator a bit since it
a special case can now always be used when an xproto::Gravity field
appears instead of just when it appears inside of the code generated for
xproto.

Related-to: https://github.com/psychon/x11rb/issues/384
Signed-off-by: Uli Schlachter <psychon@znc.in>