Remove an Option in SumOfExpr
A <sumof> expression contains a child expression in the XML. This
expression can be empty, which means that just the elements that are
iterated over are summed directly. This is equivalent to an expression
of <listelement-ref/>.

So far, no expression was represented as None and a <listelement-ref/>
was represented as Some(Expression::ListElementRef). Since both
expressions are equivalent, the Option is not really necessary.

This commit removes the Option be instead using
Expression::ListElementRef.

Signed-off-by: Uli Schlachter <psychon@znc.in>