Update doc/generated_code.md
Instead of providing the generated code inline, this links to GitHub.
These links include the commit id and thus do not break in updates.

This also makes this file hopefully easier to update: Just
search&replace the commit hash, upload to GitHub and check that all the
line numbers still match.

Signed-off-by: Uli Schlachter <psychon@znc.in>