Remove the generic arguments from RustConnection
This changes RustConnection to only support "the base case". It is not
possible to replace the underlying write/read streams with own
implementations.

The idea is that this simplifies some further development and once we
know where we end up, we bring back the generic arguments.

For now, this means that multithread_test.rs has to be removed.

Signed-off-by: Uli Schlachter <psychon@znc.in>