Make sure to wake up readers when done reading
Only one thread may read at the same time. This is arbitrated with a
mutex. Threads that do not get the mutex wait on a condition variable
until the current reader is done reading, at which point it notifies the
condition variable.

However, the notification was only sent when the reader successfully
finished reading. When some error lead to an early return, no
notification was generated. This could lead to readers hanging
indefinitely.

Fix this by notifying the condition variable in the Drop implementation
of some struct.

Signed-off-by: Uli Schlachter <psychon@znc.in>