RustConnection: Batch reads
Instead of reading just one packet from the reader whenever anything is
read, this code now "drains" the connection, meaning that it reads as
many packets as possible.

No idea about the performance implications of this (I did not measure
anything), but this is needed for some later change that I plan.

Signed-off-by: Uli Schlachter <psychon@znc.in>