BufWriteFD: Do not buffer large writes
When something tries to write more data than the buffer size, buffering
some of this data just means that the next buffer flush will do a small
write and the caller then has to write the remainder of its data.

Just writing directly to the inner writer is more efficient in this
case.

This also adds some more comments and an assertion. Also, a test needs
to be adapted to this new behaviour.

Signed-off-by: Uli Schlachter <psychon@znc.in>