Convert `Stream` into a opaque struct.
The previous `Stream` allowed users of x11rb to do directly `Stream::TcpStream(stream)`, which could lead to bugs if `stream` is not in non-blocking mode.

Instead, `Stream::from_tcp_stream` and `Stream::from_unix_stream` functions are provided, which allows creating a `Stream` while making sure it is set in non-blocking mode.