Use Cows for (most) members with lifetimes.
When serializing requests, whether or not our data has a lifetime is not that important, because we can just return a Cow from the struct's serialize function. This is the corresponding change for parsing requests. When parsing, values that are generated during parsing must be owned and not borrowed. The only exception to this is raw byte buffers which can remain untouched.

While we're here, remove unnecessary length erasures (i.e. &foo[..]) on &'input [T]s. This is only needed for fixed-length arrays.