Change return type of VoidCookie::check()
When checking whether a request without a reply succeeded, there are
three possible outcomes:
1. it succeeded
2. it failed with an X11 error
3. it failed with a ConnectionError (e.g. I/O error)

Before this commit, check() returns a Result<Option<Error>>. Case 1 is
represented as Ok(None), case 2 as as Ok(Some(err)), and case 3 as
Err(err).

This API emphasises the difference between an X11 error and a
ConnectionError. Also, it makes it harder to "bubble up" errors via ?,
because an X11 error is not an error in the Rust sense (does not
implement std's Error trait).

This commit changes check() to return Result<(), ReplyError>. Case 1 is
represented as Ok(()) and cases 2 and 3 are both an Err. This seems more
consistent.

Fixes: https://github.com/psychon/x11rb/issues/466
Signed-off-by: Uli Schlachter <psychon@znc.in>