Add a PollMode enumeration
Before this, Stream::poll() got two booleans as argument that describe
if we want to poll for readability/writability. The combination
false,false was forbidden.

This commit adds a PollMode enumeration. This makes it impossible to
represent the forbidden combination. Also, this is more readable in
callers, since one does not have to remember what the meaning of each of
the booleans is.

Signed-off-by: Uli Schlachter <psychon@znc.in>