Fix BigRequests support
The length field in requests has 16 bit. For larger requests, the big
requests extension was invented. This sets the 16 bit length field to
zero, which indicates that a new 32 bit length field comes immediately
next.

x11rb transparently uses the BigRequests extension were needed. However,
up to now, it did this incorrectly. Because of the extra 32 bit length
field, the request is now four bytes larger. Since the length field
counts the length of the request in four byte units, this means that we
need to increase the length field by one.

This commit does exactly that.

Fixes: https://github.com/psychon/x11rb/issues/480
Signed-off-by: Uli Schlachter <psychon@znc.in>