Add a 'cursor' feature
It occurred to me that there is a lot of code in src/cursor/. It is
larger than about half of the files in src/protocol/ (although quite a
bit of that size is tests). Still, cursor support is something that most
people will likely not need. Thus, it makes sense to add a feature for
it.

This commit adds a 'cursor' feature and makes it require 'render'. That
way, the cursor code is simplified a bit, because it can always assume
that render is available. Plus, no one would want 'cursor' without
'render' anyway.

Signed-off-by: Uli Schlachter <psychon@znc.in>