Add a send() method on request structures
This method allows sending a request struct directly to the X11 server.
This allows more readable code for requests with many entries, because
one can now fill the struct by name, instead of passing the arguments to
the request sending function by order (what was argument 5 to
create_window() again?).

This is now easily possible thanks to all the work done by @khuey.

Signed-off-by: Uli Schlachter <psychon@znc.in>