Make Image::put split up too large images
This function now honors the server's maximum request size. If an image
is too large to be sent in one go, it is split into multiple PutImage
requests.

This was tested by adding `let lines_per_request = 4;` after the
lines_per_request computation and making examples/display_ppm print the
number of cookies that were used. The resulting program was run under
xtrace with a non-trivial PPM.

Signed-off-by: Uli Schlachter <psychon@znc.in>