Add missing fields to X11 errors
All X11 errors have the same fields, but the xcb-proto XML still defines
them again each time. Worse, some of the definitions are missing some of
the fields.

This commit makes xcbgen-rs add the missing fields to all error
definitions. The same approach is being proposed for xcbgen (the Python
code that is part of xcb-proto) in
https://gitlab.freedesktop.org/xorg/proto/xcbproto/-/merge_requests/11

Signed-off-by: Uli Schlachter <psychon@znc.in>