Mark some enums as non-exhaustive
Only Rust 1.40 and later have #[non_exhaustive]. Still, it is a good
idea to mark enums as non-exhaustive. Thus, this commit introduces a new
features with an ugly name so that hopefully no one uses it. With this
feature, nothing changes. Without this feature, some enums are marked as
non_exhaustive.

Fixes: https://github.com/psychon/x11rb/issues/306
Signed-off-by: Uli Schlachter <psychon@znc.in>