simple_wm: Use the save set
The save set can be used by the WM so that the windows survive a WM
crash (simulated by pressing ctrl-C to abort simple_window_manager).
Without the save set, the X11 server destroys all windows created by
simple_window_manager, including their children. With the save set, the
child windows (that are in the save set) are first reparented to the
root window and survive.

When a window is closed, it is removed from the save set. This commit
also reparents the window back to the root window so that it is also not
destroyed when explicitly closing.

Signed-off-by: Uli Schlachter <psychon@znc.in>