record example: Honor X11RB_EXAMPLE_TIMEOUT
Previously, the tests on Travis would just hang. This commit makes the
test actually exist after X11RB_EXAMPLE_TIMEOUT.

This also now tests more of the record extension, namely this actually
test that the iterator stops when the right packet is sent by the server
in response to the DisableContext request.

Signed-off-by: Uli Schlachter <psychon@znc.in>