Remove some unneeded TryFrom impls from the generated code
A X11 enumeration that is represented by an u8 (because all its uses in
the protocol fit into an u8) implements TryFrom<u16> and TryFrom<u32>
before this commit. The implementation would succeed if the given value
fits into an u8 and otherwise return a ParseError. The same happens with
enumerations fitting into u16 with TryFrom<u32>.

This commit removes these impls, because I fail to see why we need them
and they seem to be unused.

Signed-off-by: Uli Schlachter <psychon@znc.in>