Improve Debug-printing of enum values
When we switched from representing enums as Rust enums to representing
them as newtypes around numbers, the derived Debug implementation
followed and instead of printing pretty names, we got just numbers. This
commit stops deriving a Debug implementation and instead one is
implemented manually.

The new Debug impl uses the alternate() flag to decide between
UPPER_CASE and CamelCase. By default, the UPPER_CASE representation that
is now also used in the source is used. With the alternate() flag,
CamelCase is used instead.

This also prints bitmasks as combinations of the individual values. For
example, ModMask::SHIFT | ModMask::LOCK actually is represented as
"SHIFT | LOCK". This was not possible before when we used a Rust enum to
represent these.

Signed-off-by: Uli Schlachter <psychon@znc.in>