Use type inference for shorter lines
We do not have to write Foo::try_parse() when we want a Foo, but we can
also just write TryParse::try_parse() and let type inference figure out
things. This leads to more characters in this example, but is actually
shorter in lots of cases.

Thanks to @khuey for suggesting (basically) this in
https://github.com/psychon/x11rb/pull/581#issuecomment-753012240.

Signed-off-by: Uli Schlachter <psychon@znc.in>