Fix CI
The newest clippy beta brought a new lint that complains about some
names, e.g. GC should be Gc. This is not really helpful for us since
these are names that come from the XML description and we really should
not be too different from the XML "just because". Thus, silence this
lint on CI.

Fixes: https://github.com/psychon/x11rb/issues/596
Signed-off-by: Uli Schlachter <psychon@znc.in>