Fix lots of "Question mark operator is useless here"
The newest clippy beta added a new lint that warns about useless use of
question mark operators. This commit changes the code generator to fix
this.

There are still other warnings from this lint in x11rb, but I did not
want to hide this in such a large commit.

Signed-off-by: Uli Schlachter <psychon@znc.in>