Add ClientMessageEvent::new()
This commit adds a helper function that simplifies creation of
ClientMessageEvent by always setting sequence to 0 and response_type to
CLIENT_MESSAGE_EVENT.

Additionally, the atom is specified via Into<Atom>, thus allowing to use
AtomEnum directly. The same applies to the data member which is given as
Into<ClientMessageData>. Thus, it is possible to directly pass in
fixed-size arrays.

Fixes: https://github.com/psychon/x11rb/issues/592
Signed-off-by: Uli Schlachter <psychon@znc.in>