Add request name to X11Error
This adds a function to get the name of a request from the name of the
extension and the major and minor opcode of a request.

This was tested with the xeyes example by replacing "window_id" with
"window_id - 1" in a request. The debug output for the resulting X11
protocol error was:

Unknown error X11Error { error_kind: Window, error_code: 3, sequence: 19, bad_value: 67108863, minor_opcode: 2, major_opcode: 129, extension: Some("SHAPE"), request: Some("Mask") }

Fixes: https://github.com/psychon/x11rb/issues/614
Signed-off-by: Uli Schlachter <psychon@znc.in>