Simplify request serialisation function arguments
Before this change, a request struct had a serialize() method that has a
generic argument for a RequestConnection. The current extension's major
opcode was requested from this connection. For xproto, this extra
argument was just ignored.

This commit changes the code so that instead the argument is a
non-generic major_opcode, which is all that is needed from the
connection. The "get the opcode from the connection"-part is moved into
the callers instead.

Thanks to this change, serialize() becomes infallible and no longer
needs to return a Result.

Besides a better/simpler API, another benefit of this commit is that it
is a net reduction in lines of code: 1558 lines of code are removed.

Signed-off-by: Uli Schlachter <psychon@znc.in>