Remove some more unnecessary unwrap()s
This does the same thing as the previous commit (replacing TryFrom with
From), but in another place. This again saves some unwrap()s.

Note that the generated code most likely now uses From to convert from a
type T to the type T itself, aka a no-op. This is still done on purpose
to make sure that we end up with the correct type. Otherwise, it becomes
to easy to have the wrong type here and then serialise a too large / too
small field.

This commit also made the decision on what type to use a bit more
clever. The code now looks through unary and binary operations and still
picks a type other than u32 if all the sub-expressions have this type.
This is necessary to make XKB's SelectEvents work.

Also, there was a bug in expr_to_str() where it still used u32 for
unary/binary operations even when another type was requested. This
commit fixes that as well.

Helps-with: https://github.com/psychon/x11rb/issues/178
Signed-off-by: Uli Schlachter <psychon@znc.in>