Be safe when deducing deducible fields
Now that switch_expr()'s return type is "complicated", we should make
sure that we really get the expected return type from it. Thus, this
removes an optimisation that assumed that switch_expr() always returned
u32.

Instead, we now emit a type annotation that "documents" the expected
type. If switch_expr() were to return something else, this would thus
fail to compile.

Since this also removes the calls to From::from, this fixes the
following kind clippy warning:

error: useless conversion to the same type: `u16`
    --> src/protocol/xproto.rs:7870:26
     |
7870 |         let value_mask = u16::from(self.value_list.switch_expr());
     |                          ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ help: consider removing `u16::from()`: `self.value_list.switch_expr()`
     |
     = note: `-D clippy::useless-conversion` implied by `-D warnings`
     = help: for further information visit https://rust-lang.github.io/rust-clippy/master/index.html#useless_conversion

Signed-off-by: Uli Schlachter <psychon@znc.in>