Add traits for request serialisation
This commit changes the Request trait to have a new EXTENSION_NAME
associated constant and a serialize() method. This allows any user of
the trait to serialize a Request and send it to an X11 server.

This also adds VoidRequest, ReplyRequest, ReplyFDsRequest marker traits
that allow to figure out what kind of reply to expect from sending such
a request.

The code generator is extended to implement these changes.

Closes: https://github.com/psychon/x11rb/issues/656
Signed-off-by: Uli Schlachter <psychon@znc.in>