Refactor: Do not do I/O in IdAllocator
Previously, the IdAllocator needed an X11 connection and did I/O on this
connection if it ran out of XIDs. This commit moves that I/O to the
caller and the IdAllocator just indicates that it ran out of XIDs. A new
function update_xid_range() is added that gets the resulting
GetXIDRangeReply and updates the state accordingly.

Moving things around in this way makes it possible to add the
IdAllocator to the public API since it is reusable. This also allows to
add a new test case invalid_update_arg() for a case that was previously
not testable.

Thanks a lot to @notgull for giving me the idea.

Signed-off-by: Uli Schlachter <psychon@znc.in>