Split out a x11rb-protocol crate
This commit splits the code into two crates:

"x11rb-protocol" contains the protocol definitions and allows parsing
and serialising things. It does not do any actual I/O.

"x11rb" uses x11rb-protocol and implements an X11 client based on this.

I tried to do as few changes as possible here. Code moves around, but
the structure of the code is not supposed to change.

There is one place where this did not work out:

Previously, e.g. CreateWindowRequest had an associated function send().
This function send the request that it is called on via a given X11
connection. Since this does I/O, this functionality does not belong into
the x11rb-protocol crate. Since it is an associated function, it cannot
easily be provided by x11rb.

For now, I worked around this problem by turning send() into a "free"
function. The resulting name collision was dealt by renaming it to e.g.
send_create_window().

The result of all this is a huge commit. This is still a mess and things
should be cleaned up further in the future, but I didn't want to make
this commit even larger. This is the first version that I managed to get
to compile.

Signed-off-by: Uli Schlachter <psychon@znc.in>