Add a mechanism to send requests based on traits
Commit 105c0fc3af added three new traits that allow to send a request
simply based on a trait. This commit extends our RequestConnection trait
with functions that use this capability to actually send a request.

For a request which has a reply without FDs, the reply type has to
implement TryParse, but before this commit, the trait only provided
TryParseFd. Thus, this commit moves the "Reply" associated type around
so that only for ReplyFDsRequest it has to implement TryParseFd.

Signed-off-by: Uli Schlachter <psychon@znc.in>