Remove "Reply" field from "Request" trait
The previous commit added a "Reply" associated type to the
"ReplyRequest" trait. Now, both "Request" and "ReplyRequest" have such
an associated type with a different type. To clean that up, this commit
moves Request::Reply to ReplyRequest::Reply. The code generator is
patched up accordingly.

Additionally, Request::parse_reply() is removed. Instead, the same
functionality is implemented in x11rb-protocol/src/protocol/mod.rs
locally (and not "pub"lically), which is the only place where this was
used before.

Signed-off-by: Uli Schlachter <psychon@znc.in>