Move x11rb to x11rb/ folder
Making this work on CI required some changes to the GitHub workflow. As
far as I can tell, this dance with $ALL_FEATURES was added "way back" to
work around libloading not working on our MSRV back then. Since this
problem is long gone, this commit cleans that up and uses
--all-features instead.

Signed-off-by: Uli Schlachter <psychon@znc.in>