Document abstract unix socket use
We already have ConnectAddress::Socket that means "the given file name
is a unix socket to connect to". Abstract unix sockets work the same
name, only that the file should be looked up in the abstract namespace
instead of the filesystem.

Thus, this commit changes the docs of x11rb-protocol accordingly and
moves the "TODO" for implementing this from x11rb-protocol to x11rb.

Additionally, this marks ConnectAddress as #[non_exhaustive] so that we
may add new cases in the future. One might think "this will never happen
/ be necessary", but here is a merge request trying to add a new case to
libxcb:
https://gitlab.freedesktop.org/xorg/lib/libxcb/-/merge_requests/27

Signed-off-by: Uli Schlachter <psychon@znc.in>