Add support for connecting over abstract unix sockets
This adds support for connecting over abstract unix sockets. The
connection is established via functions provided by nix. The resulting
socket is wrapped in a RawFdContainer so that it is automatically closed
on drop.

All the actual I/O already uses as_raw_fd() to access the RawFd and work
directly with that. Thus, the existing code for this just works.

In the future, when there is a safe way to construct a UnixStream from a
RawFd, the AbstractUnixStream case that this commit adds should be be
removed again (e.g. nix could add wrappers using OwnedFd, when that type
becomes stable).

Signed-off-by: Uli Schlachter <psychon@znc.in>